+++ packages/rax-miniapp-runtime-webpack-plugin/src/platforms/ali.js
@@ -745,6 +745,27 @@ const ContactButton = {
   }
 };
 
+const Lifestyle = {
+  props: {
+    publicId: '',
+    memo: ''
+  },
+  events: {
+    Follow: ''
+  }
+};
+
+const LifeFollow = {
+  props: {
+    sceneId: '',
+    checkFollow: ''
+  },
+  events: {
+    CheckFollow: '',
+    Close: ''
+  }
+};
+
 exports.internalComponents = {
   View,
   CatchView,
@@ -792,6 +813,8 @@ exports.internalComponents = {
   LivePlayer,
   LivePusher,
   ContactButton,
+  Lifestyle,
+  LifeFollow,
   HElement,
   NoTouchHElement,
   CatchHElement,
@@ -856,7 +879,9 @@ exports.voidChildrenElements = new Set([
   'LivePlayer',
   'LivePusher',
   'ContactButton',
-  'OpenAvatar'
+  'OpenAvatar',
+  'Lifestyle',
+  'LifeFollow'
 ]);