+++ packages/build-plugin-rax-miniapp-faas/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.2
+
+### Changed
+
+- update miniapp-builder-shared to v3.0.0
+
 ## v1.0.1
 
 - Fix: can't copy miniapp-native to correct build dir